[S07P22D208-212 | AOS] feat: JoinProviderFragment

+++ aos/taxi/app/src/main/java/com/example/taxi/data/repository/UserInfoRepositoryImpl.kt
@@ -125,16 +125,33 @@ class UserInfoRepositoryImpl(private val database: FirebaseFirestore) : UserInfo
     }
 
     override fun deleteImage(result: (UiState<String>) -> Unit){
-        var storage = FirebaseStorage.getInstance()
-        var imgFileName = ApplicationClass.prefs.userSeq.toString() + ".png"
+        if(ApplicationClass.prefs.profileImage != ""){
+            var storage = FirebaseStorage.getInstance()
+            var imgFileName = ApplicationClass.prefs.userSeq.toString() + ".png"
 
-        storage.getReference().child("user_profiles").child(imgFileName)
-            .delete()
-            .addOnSuccessListener {
-                Log.d("addImageUpLoad", "Image has been uploaded successfully")
-            }.addOnFailureListener{
-                Log.d("addImageUpLoad", "Image has been uploaded fail")
+            storage.getReference().child("user_profiles").child(imgFileName)
+                .delete()
+                .addOnSuccessListener {
+                    Log.d("addImageUpLoad", "Image has been uploaded successfully")
+                }.addOnFailureListener{
+                    Log.d("addImageUpLoad", "Image has been uploaded fail")
+                }
+
+            if(ApplicationClass.prefs.isEachProvider == true){
+                if(ApplicationClass.prefs.carImage != ""){
+                    var storage = FirebaseStorage.getInstance()
+                    var imgFileName = ApplicationClass.prefs.providerId.toString() + ".png"
+
+                    storage.getReference().child("provider_car_profiles").child(imgFileName)
+                        .delete()
+                        .addOnSuccessListener {
+                            Log.d("addImageUpLoad", "Image has been uploaded successfully")
+                        }.addOnFailureListener{
+                            Log.d("addImageUpLoad", "Image has been uploaded fail")
+                        }
+                }
             }
+        }
     }
     override fun getBoardedTaxiList(result: (UiState<BoardedTaxiList>) -> Unit) {
         database.collection(FireStoreCollection.BOARDEDTAXILIST).document(ApplicationClass.userId)
